+++ Engine/Renderer/GLES3/GLES3CommandDrawElements.h
@@ -31,6 +31,6 @@ public:
 private:
 	uint32_t m_mode;
 	uint32_t m_type;
-	uint32_t m_offset;
+	uintptr_t m_offset;
 	int m_count;
 };
