+++ Engine/Renderer/GLES3/GLES3CommandDrawIndirect.h
@@ -30,5 +30,5 @@ public:
 private:
 	uint32_t m_mode;
 	uint32_t m_type;
-	uint32_t m_offset;
+	uintptr_t m_offset;
 };
